feat: Add @radix-ui/react-select dependency and implement updateMemberAction
This commit adds the "@radix-ui/react-select" dependency to the package.json file. It also implements the "updateMemberAction" function in the "update-member.ts" file, which allows updating the role of a member in an organization. The "updateMemberAction" function is used in the "update-member-role-select.tsx" file to update the role of a member using a select component.