feat: Add dayjs library for date manipulation and formatting
This commit adds the dayjs library as a new dependency to the project. Dayjs is a lightweight JavaScript library for date manipulation and formatting, providing a simpler and more intuitive API compared to other date libraries. This addition will enable the use of relative time formatting in the project's project list component, improving the user experience by displaying more human-readable timestamps.